requirements.txt
setuptools
tornado
pymysql
cx_Oracle
psycopg2-binary
Cython
pymssql
pyinstaller
